refactor: move emitAlias function
